use shared BlocksGatway
