+++ examples/gateway/proxy/main_test.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/ipfs/go-libipfs/gateway"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -22,7 +23,7 @@ func newProxyGateway(t *testing.T, rs *httptest.Server) *httptest.Server {
 	blockService := blockservice.New(blockStore, offline.Exchange(blockStore))
 	routing := newProxyRouting(rs.URL, nil)
 
-	gateway, err := common.NewBlocksGateway(blockService, routing)
+	gateway, err := gateway.NewBlocksGateway(blockService, gateway.WithValueStore(routing))
 	if err != nil {
 		t.Error(err)
 	}
